update: wrappers
1. moved flow wrappers to separate class
2. added safe call wrapper
